separated reaction code into 2 files
