+++ bruinbook-frontend/src/components/Reaction.css
@@ -1,0 +1,6 @@
+  .button {
+    background: white;
+    border-radius: 10%;
+    height: 60px;
+    border:0cm
+}
