+++ bruinbook-frontend/src/components/ReactionBar.css
@@ -5,12 +5,6 @@
     
   }
 
-  .button {
-      background: white;
-      border-radius: 10%;
-      height: 60px;
-      border:0cm
-  }
 
   .bar-container{
       background: white;
